First Commit
This program allows the user to enter and search for strings.   When
strings are added to the tree, they should be wrapped inside a node
object that holds the string, the frequency with which that string has
been added to the tree, and references to two other nodes.  The strings
are stored in a balanced Binary Tree